# [inductor] Compile time optimizations