--- a/benchmarks/common.py
+++ b/benchmarks/common.py
@@ -116,6 +116,7 @@ CI_SKIP_TRAINING = [
     "T5Small",
     "XGLMForCausalLM",
     "XLNetLMHeadModel",
+    "PegasusForCausalLM",
 ]