@jansel 2022-08-29
Note this adds a CI skip for PegasusForCausalLM.  I don't think the following error is related to this PR:
```
cuda train PegasusForCausalLM                 ERROR:root:unhandled error
Traceback (most recent call last):
  File "/home/circleci/project/benchmarks/common.py", line 1139, in run_one_model
    new_result = optimized_model_iter_fn(model, example_inputs)
  File "/home/circleci/project/torchdynamo/eval_frame.py", line 154, in _fn
    return fn(*args, **kwargs)
  File "benchmarks/huggingface.py", line 448, in forward_and_backward_pass
    def forward_and_backward_pass(self, mod, inputs, collect_outputs=True):
  File "benchmarks/huggingface.py", line 448, in forward_and_backward_pass
    def forward_and_backward_pass(self, mod, inputs, collect_outputs=True):
  File "benchmarks/huggingface.py", line 454, in forward_and_backward_pass
    self.grad_scaler.scale(loss).backward()
  File "/home/circleci/project/env/lib/python3.8/site-packages/functorch/_src/monkey_patching.py", line 77, in _backward
    return _old_backward(*args, **kwargs)
  File "/home/circleci/project/env/lib/python3.8/site-packages/torch/_tensor.py", line 484, in backward
    torch.autograd.backward(
  File "/home/circleci/project/env/lib/python3.8/site-packages/torch/autograd/__init__.py", line 191, in backward
    Variable._execution_engine.run_backward(  # Calls into the C++ engine to run the backward pass
  File "/home/circleci/project/env/lib/python3.8/site-packages/torch/autograd/function.py", line 267, in apply
    return user_fn(self, *args)
  File "/home/circleci/project/env/lib/python3.8/site-packages/functorch/_src/aot_autograd.py", line 365, in backward
    CompiledFunction.compiled_bw = aot_config.bw_compiler(
  File "/home/circleci/project/torchdynamo/optimizations/backends.py", line 542, in _wrapped_bw_compiler
    return torchdynamo.disable(bw_compiler(*args, **kwargs))
  File "/home/circleci/project/torchinductor/compile_fx.py", line 176, in bw_compiler
    return compile_fx_inner(
  File "/home/circleci/project/torchdynamo/debug_utils.py", line 253, in debug_wrapper
    raise e
  File "/home/circleci/project/torchdynamo/debug_utils.py", line 241, in debug_wrapper
    compiled_fn = compiler(gm, example_inputs, **kwargs)
  File "/home/circleci/project/torchinductor/debug.py", line 182, in inner
    return fn(*args, **kwargs)
  File "/home/circleci/project/torchinductor/compile_fx.py", line 60, in compile_fx_inner
    compiled_fn = cudagraphify(
  File "/home/circleci/project/torchinductor/compile_fx.py", line 102, in cudagraphify
    model(*inputs)
  File "/tmp/torchinductor_circleci/2a/c2amnj4bvgezefiyv4h23vy34iycx3kg377uhapp6obobswi5h5p.py", line 1051, in call
    kernel0[grid(128)](add_tensor, div_default, reciprocal_default, primals_1, primals_2, addmm_default_3, addmm_default_5, div_default_2, reciprocal_default_3, primals_17, primals_18, addmm_default_9, div_default_3, reciprocal_default_4, addmm_default_11, primals_27, primals_28, div_default_4, reciprocal_default_6, primals_33, primals_34, addmm_default_15, addmm_default_17, div_default_6, reciprocal_default_9, primals_49, primals_50, addmm_default_21, div_default_7, reciprocal_default_10, addmm_default_23, primals_59, primals_60, div_default_8, reciprocal_default_12, primals_65, primals_66, addmm_default_27, addmm_default_29, div_default_10, reciprocal_default_15, primals_81, primals_82, addmm_default_33, div_default_11, reciprocal_default_16, addmm_default_35, primals_91, primals_92, div_default_12, reciprocal_default_18, primals_97, primals_98, addmm_default_39, addmm_default_41, div_default_14, reciprocal_default_21, primals_113, primals_114, addmm_default_45, div_default_15, reciprocal_default_22, addmm_default_47, primals_123, primals_124, div_default_16, reciprocal_default_24, primals_129, primals_130, addmm_default_51, addmm_default_53, div_default_18, reciprocal_default_27, primals_145, primals_146, addmm_default_57, div_default_19, reciprocal_default_28, addmm_default_59, primals_155, primals_156, div_default_20, reciprocal_default_30, primals_161, primals_162, addmm_default_63, addmm_default_65, div_default_22, reciprocal_default_33, primals_177, primals_178, addmm_default_69, div_default_23, reciprocal_default_34, addmm_default_71, div_default_24, reciprocal_default_36, div_default_21, reciprocal_default_31, primals_171, primals_172, primals_187, primals_188, tangents_1, primals_193, buf0, buf1, buf2, buf3, buf4, buf440, buf5, buf6, buf7, buf8, buf9, buf360, buf10, buf11, buf12, buf13, buf14, buf280, buf15, buf16, buf17, buf18, buf19, buf200, buf20, buf21, buf22, buf23, buf24, buf120, buf25, buf26, buf27, buf28, buf29, buf80, buf40, buf32, 128, 1024)
  File "/home/circleci/project/env/lib/python3.8/site-packages/triton/code_gen.py", line 999, in __call__
    return self.kernel(*wargs, **kwargs, grid=self.grid)
  File "/home/circleci/project/env/lib/python3.8/site-packages/triton/code_gen.py", line 1073, in __call__
    timings = {config: self._bench(*args, config=config, **kwargs)
  File "/home/circleci/project/env/lib/python3.8/site-packages/triton/code_gen.py", line 1073, in <dictcomp>
    timings = {config: self._bench(*args, config=config, **kwargs)
  File "/home/circleci/project/torchinductor/triton_ops/autotune.py", line 28, in _bench
    return super()._bench(*args, config=config, **kwargs)
  File "/home/circleci/project/env/lib/python3.8/site-packages/triton/code_gen.py", line 1054, in _bench
    return triton.testing.do_bench(kernel_call)
  File "/home/circleci/project/env/lib/python3.8/site-packages/triton/testing.py", line 138, in do_bench
    torch.cuda.synchronize()
  File "/home/circleci/project/env/lib/python3.8/site-packages/torch/cuda/__init__.py", line 501, in synchronize
    return torch._C._cuda_synchronize()
RuntimeError: CUDA error: misaligned address
CUDA kernel errors might be asynchronously reported at some other API call,so the stacktrace below might be incorrect.
For debugging consider passing CUDA_LAUNCH_BLOCKING=1.
ERROR
CUDA error: misaligned address
CUDA kernel errors might be asynchronously reported at some other API call,so the stacktrace below might be incorrect.
For debugging consider passing CUDA_LAUNCH_BLOCKING=1.
ERROR
```
Likely just uncovered by different fusions happening.