@Chillee 2022-08-29
So... confirmed that `aggressive_fusions` doesn't actually improve perf? :P Even in low-batch size settings?