@jansel 2022-08-29
> So... confirmed that `aggressive_fusions` doesn't actually improve perf? :P Even in low-batch size settings?

Overall it is a pretty small effect.  Though I suspect cudagraphs is also helping here.  If we use only 1 stream it might matter more.